+++ react/typings/global.d.ts
@@ -17,10 +17,7 @@ declare interface Pickup {
 
 declare type CallToAction = 'popover-button' | 'popover-input' | 'modal'
 
-declare type ShippingSelection =
-  | 'only-pickup'
-  | 'delivery-and-pickup'
-  | 'only-delivery'
+declare type ShippingSelection = 'only-pickup' | 'delivery-and-pickup'
 
 declare type Mode = 'default' | 'icon'