+++ react/context/useShippingOption.ts
@@ -382,8 +382,8 @@ export const useShippingOption = () => {
           })
         )
 
-        setActionInterruptedByCartValidation(() => () =>
-          submitZipcode(zipcodeSelected, reload)
+        setActionInterruptedByCartValidation(
+          () => () => submitZipcode(zipcodeSelected, reload)
         )
 
         break
@@ -465,8 +465,8 @@ export const useShippingOption = () => {
           )
         )
 
-        setActionInterruptedByCartValidation(() => () =>
-          selectDeliveryShippingOption()
+        setActionInterruptedByCartValidation(
+          () => () => selectDeliveryShippingOption()
         )
 
         break