+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Unselect the pickup point in the global filter.
+
 ## [1.9.0] - 2025-10-09
 
 ### Added