+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Not forwarding cookies to `getCatalogCount` anymore
+- Refresh page after `zipcode update` at `segment cookie`
+
 ## [1.10.0] - 2025-10-28
 
 ### Added